Add Message related contents to VoiceChannel
