+++ src/main/java/snw/kookbc/impl/HttpAPIImpl.java
@@ -31,14 +31,14 @@ import snw.jkook.entity.Guild;
 import snw.jkook.entity.User;
 import snw.jkook.entity.channel.Category;
 import snw.jkook.entity.channel.Channel;
-import snw.jkook.entity.channel.TextChannel;
+import snw.jkook.entity.channel.NonCategoryChannel;
+import snw.jkook.message.ChannelMessage;
 import snw.jkook.message.PrivateMessage;
-import snw.jkook.message.TextChannelMessage;
 import snw.jkook.message.component.BaseComponent;
 import snw.jkook.util.PageIterator;
 import snw.jkook.util.Validate;
+import snw.kookbc.impl.message.ChannelMessageImpl;
 import snw.kookbc.impl.message.PrivateMessageImpl;
-import snw.kookbc.impl.message.TextChannelMessageImpl;
 import snw.kookbc.impl.network.HttpAPIRoute;
 import snw.kookbc.impl.network.exceptions.BadResponseException;
 import snw.kookbc.impl.pageiter.GameIterator;
@@ -224,7 +224,7 @@ public class HttpAPIImpl implements HttpAPI {
     }
 
     @Override
-    public TextChannelMessage getTextChannelMessage(String id) throws NoSuchElementException {
+    public ChannelMessage getChannelMessage(String id) throws NoSuchElementException {
         final JsonObject object;
         try {
             object = client.getNetworkClient()
@@ -241,14 +241,14 @@ public class HttpAPIImpl implements HttpAPI {
         User sender = client.getStorage().getUser(get(rawSender, "id").getAsString(), rawSender);
         final BaseComponent component = client.getMessageBuilder().buildComponent(object);
         long timeStamp = get(object, "create_at").getAsLong();
-        TextChannelMessage quote = null;
+        ChannelMessage quote = null;
         if (has(object, "quote")) {
             final JsonObject rawQuote = get(object, "quote").getAsJsonObject();
             final String quoteId = get(rawQuote, "id").getAsString();
-            quote = getTextChannelMessage(quoteId);
+            quote = getChannelMessage(quoteId);
         }
-        final TextChannel channel = (TextChannel) getChannel(get(object, "channel_id").getAsString());
-        return new TextChannelMessageImpl(client, id, sender, component, timeStamp, quote, channel);
+        final NonCategoryChannel channel = (NonCategoryChannel) getChannel(get(object, "channel_id").getAsString());
+        return new ChannelMessageImpl(client, id, sender, component, timeStamp, quote, channel);
     }
 
     @Override
