+++ src/main/java/snw/kookbc/impl/message/VoiceChannelMessageImpl.java
@@ -1,0 +1,23 @@
+package snw.kookbc.impl.message;
+
+import snw.jkook.entity.User;
+import snw.jkook.entity.channel.VoiceChannel;
+import snw.jkook.message.Message;
+import snw.jkook.message.VoiceChannelMessage;
+import snw.jkook.message.component.BaseComponent;
+import snw.kookbc.impl.KBCClient;
+
+public class VoiceChannelMessageImpl extends ChannelMessageImpl implements VoiceChannelMessage {
+
+    private final VoiceChannel channel;
+
+    public VoiceChannelMessageImpl(KBCClient client, String id, User user, BaseComponent component, long timeStamp, Message quote, VoiceChannel channel) {
+        super(client, id, user, component, timeStamp, quote, channel);
+        this.channel = channel;
+    }
+
+    @Override
+    public VoiceChannel getChannel() {
+        return channel;
+    }
+}
